style(LODGroupTweaks): :art: align patch with file naming rules
