+++ .editorconfig
@@ -52,6 +52,7 @@ dotnet_analyzer_diagnostic.category-Roslynator.severity = warning
 
 # Specific overrides for .NET Framework compatibility
 dotnet_diagnostic.CA2007.severity = none     # ConfigureAwait for .NET Framework is less relevant
+dotnet_diagnostic.CA1303.severity = none     # mod logs are not localized
 
 # IDE-specific overrides
 dotnet_diagnostic.IDE0058.severity = none    # Allow implicit discard of expression values
@@ -70,6 +71,7 @@ dotnet_diagnostic.CA1707.severity = none  # test classes often don't follow nami
 # StyleCop: configure per rule (base off)
 dotnet_analyzer_diagnostic.category-StyleCop.severity = none
 dotnet_diagnostic.SA1649.severity = error
+dotnet_diagnostic.SA1111.severity = warning
 
 # StyleCop exclusions (repo policy)
 # - File header/element docs are not required; keep code concise
@@ -87,7 +89,6 @@ dotnet_diagnostic.SA1107.severity = none  # multiple statements on one line
 dotnet_diagnostic.SA1202.severity = none  # member order rules
 dotnet_diagnostic.SA1311.severity = none  # static readonly fields casing
 dotnet_diagnostic.SA1009.severity = none  # closing parenthesis spacing
-dotnet_diagnostic.SA1111.severity = none  # closing parenthesis should be on line of last parameter
 dotnet_diagnostic.SA1512.severity = none  # single-line comments should not be followed by blank line
 dotnet_diagnostic.SA1402.severity = none  # file may only contain a single type
 dotnet_diagnostic.SA1516.severity = none  # elements should be separated by blank line
