+++ FluxLoopTweaks/AsyncWhile_RunAsync_Patch.cs
@@ -1,4 +1,5 @@
 using System.Diagnostics;
+using System.Diagnostics.CodeAnalysis;
 using System.Runtime.CompilerServices;
 using System.Threading.Tasks;
 using FrooxEngine.ProtoFlux;
@@ -13,54 +14,11 @@ namespace EsnyaTweaks.FluxLoopTweaks;
 [HarmonyPatch(typeof(AsyncWhile), "RunAsync")]
 internal static class AsyncWhile_RunAsync_Patch
 {
-    [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private static async Task<IOperation> RunAsync(
-        AsyncWhile instance,
-        FrooxEngineContext context,
-        ValueInput<bool> Condition,
-        AsyncCall LoopStart,
-        AsyncCall LoopIteration,
-        Continuation LoopEnd,
-        int timeout
-    )
-    {
-        await LoopStart.ExecuteAsync(context).ConfigureAwait(true);
-
-        var previousTick = context.Engine.UpdateTick;
-        var stopwatch = Stopwatch.StartNew();
-
-        while (Condition.Evaluate(context, defaultValue: false))
-        {
-            var currentTick = context.Engine.UpdateTick;
-            if (currentTick != previousTick)
-            {
-                stopwatch.Restart();
-            }
-            else if (stopwatch.ElapsedMilliseconds > timeout || context.AbortExecution)
-            {
-                ResoniteMod.Warn(
-                    $"AsyncWhile Timedout: {instance} ({stopwatch.ElapsedMilliseconds}ms)"
-                );
-                context.AbortExecution = true;
-                throw new ExecutionAbortedException(
-                    instance.Runtime as IExecutionRuntime,
-                    instance,
-                    LoopIteration.Target,
-                    isAsync: true
-                );
-            }
-            previousTick = currentTick;
-
-            await LoopIteration.ExecuteAsync(context).ConfigureAwait(true);
-        }
-        return LoopEnd.Target;
-    }
-
+    [SuppressMessage("Style", "SA1313", Justification = "Harmony magic parameters")]
     internal static bool Prefix(
         AsyncWhile __instance,
         ExecutionContext context,
-        ref Task<IOperation> __result
-    )
+        ref Task<IOperation> __result)
     {
         if (context is not FrooxEngineContext frooxEngineContext)
         {
@@ -74,8 +32,48 @@ internal static class AsyncWhile_RunAsync_Patch
             __instance.LoopStart,
             __instance.LoopIteration,
             __instance.LoopEnd,
-            FluxLoopTweaksMod.TimeoutMs
-        );
+            FluxLoopTweaksMod.TimeoutMs);
         return false;
+    }
+
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    private static async Task<IOperation> RunAsync(
+        AsyncWhile instance,
+        FrooxEngineContext context,
+        ValueInput<bool> condition,
+        AsyncCall loopStart,
+        AsyncCall loopIteration,
+        Continuation loopEnd,
+        int timeout)
+    {
+        await loopStart.ExecuteAsync(context).ConfigureAwait(true);
+
+        var previousTick = context.Engine.UpdateTick;
+        var stopwatch = Stopwatch.StartNew();
+
+        while (condition.Evaluate(context, defaultValue: false))
+        {
+            var currentTick = context.Engine.UpdateTick;
+            if (currentTick != previousTick)
+            {
+                stopwatch.Restart();
+            }
+            else if (stopwatch.ElapsedMilliseconds > timeout || context.AbortExecution)
+            {
+                ResoniteMod.Warn(
+                    $"AsyncWhile Timedout: {instance} ({stopwatch.ElapsedMilliseconds}ms)");
+                context.AbortExecution = true;
+                throw new ExecutionAbortedException(
+                    instance.Runtime as IExecutionRuntime,
+                    instance,
+                    loopIteration.Target,
+                    isAsync: true);
+            }
+            previousTick = currentTick;
+
+            await loopIteration.ExecuteAsync(context).ConfigureAwait(true);
+        }
+
+        return loopEnd.Target;
     }
 }
