+++ InventoryUITweaks.Tests/ReflectionCacheTests.cs
@@ -1,5 +1,4 @@
 using Xunit;
-using FluentAssertions;
 
 namespace EsnyaTweaks.InventoryUITweaks.Tests;
 
@@ -8,6 +7,6 @@ public class ReflectionCacheTests
     [Fact]
     public void DirectoryField_ShouldExist()
     {
-        Assert.NotNull(ReflectionCache.InventoryItemUI_Directory);
+        Assert.NotNull(ReflectionCache.InventoryItemUIDirectory);
     }
 }
