+++ InventoryUITweaks/InventoryUITweaksMod.cs
@@ -14,35 +14,34 @@ namespace EsnyaTweaks.InventoryUITweaks;
 /// <inheritdoc/>
 public class InventoryUITweaksMod : EsnyaResoniteMod
 {
-    private static Assembly ThisAssembly => typeof(InventoryUITweaksMod).Assembly;
-
     internal static string HarmonyId => $"com.nekometer.esnya.{ThisAssembly.GetName()}";
 
-    private static readonly Harmony harmony = new(HarmonyId);
+    private static Harmony Harmony { get; } = new(HarmonyId);
+
+    private static Assembly ThisAssembly => typeof(InventoryUITweaksMod).Assembly;
+
+#if DEBUG
+    /// <summary>Unpatches Harmony before hot reload.</summary>
+    public static void BeforeHotReload()
+    {
+        Harmony.UnpatchAll(HarmonyId);
+    }
+
+    /// <summary>Reapplies Harmony patches after hot reload.</summary>
+    /// <param name="mod">Unused reference to the reloaded mod.</param>
+    public static void OnHotReload(ResoniteMod mod)
+    {
+        _ = mod;
+        Harmony.PatchAll();
+    }
+#endif
 
     /// <inheritdoc/>
     public override void OnEngineInit()
     {
-        harmony.PatchAll();
-
+        Harmony.PatchAll();
 #if DEBUG
         HotReloader.RegisterForHotReload(this);
 #endif
     }
-
-#if DEBUG
-    /// <inheritdoc/>
-    public static void BeforeHotReload()
-    {
-        harmony.UnpatchAll(HarmonyId);
-    }
-
-    /// <inheritdoc/>
-    [System.Diagnostics.CodeAnalysis.SuppressMessage("Style", "IDE0060:Remove unused parameter", Justification = "Hot reload signature")]
-    public static void OnHotReload(ResoniteMod modInstance)
-    {
-        harmony.PatchAll();
-    }
-#endif
-
 }
